requirements.txt
easydict==1.9
numpy==1.23.1
pandas==1.4.3
PyYAML==6.0
scikit_learn==1.1.2
